Refactor and test
- Extract schemata for donorPayments and paymentRequests
- Write test for reimbursementReply
