+++ src/lib/airtable/tables/paymentRequests.js
@@ -1,5 +1,5 @@
 const { merge } = require("lodash");
-const { paymentsAirbase } = require("~airtable/bases");
+const { paymentRequestsTable, fields } = require("./paymentRequestsSchema");
 
 // `update` should look like:
 // {
@@ -106,55 +106,3 @@ exports.deletePaymentRequest = async (record) => {
     console.error(`Error while deleting payment request ${e}`);
   }
 };
-
-// ==================================================================
-// Schema
-// ==================================================================
-
-const paymentRequestsTableName = (exports.paymentRequestsTableName =
-  "PaymentRequests");
-const paymentRequestsTable = (exports.paymentRequestsTable = paymentsAirbase(
-  paymentRequestsTableName
-));
-const fields = (exports.paymentRequestsFields = {
-  id: "ID",
-  amount: "Amount",
-  receipts: "Receipts",
-  donation: "Donation",
-  reimbursementAmount: "ReimbursementAmount",
-  requestCode: "RequestCode",
-  donorPayments: "DonorPayments",
-  isPaid: "IsPaid",
-  created: "Created",
-  balance: "Balance",
-  lastModified: "Last Modified",
-  venmoId: "VenmoID",
-  paypalId: "PaypalID",
-  cashAppId: "CashAppID",
-  firstName: "FirstName",
-  slackMessage: "SlackMessage",
-  slackThreadId: "SlackThreadId",
-  adminNotificationSent: "AdminNotificationSent",
-  adminDenialMessage: "AdminDenialMessage",
-  phone: "Phone",
-  approval: "Approval",
-  approval_options: {
-    approved: "Approved",
-    denied: "Denied",
-  },
-  paidAmount: "PaidAmount",
-  meta: "Meta",
-  lastProcessed: "Last Processed",
-  type: "Type",
-  type_options: {
-    preimbursement: "Preimbursement",
-    reimbursement: "Reimbursement",
-    directAid: "Direct Aid",
-  },
-});
-exports.paymentRequestsSensitiveFields = [
-  fields.phone,
-  fields.venmoId,
-  fields.paypalId,
-  fields.cashAppId,
-];
