+++ src/slackapp/event-listeners/reimbursementReply.js
@@ -1,14 +1,11 @@
 const slackapi = require("~slack/webApi");
 const { findChannelByName } = require("~slack/channels");
 const { REIMBURSEMENT_CHANNEL } = require("~slack/constants");
-const {
-  findPaymentRequestBySlackThreadId,
-  paymentRequestsFields,
-} = require("~airtable/tables/paymentRequests");
-const {
-  createDonorPayment,
-  donorPaymentsFields,
-} = require("~airtable/tables/donorPayments");
+const { findPaymentRequestBySlackThreadId } = require("~airtable/tables/paymentRequests");
+const { paymentRequestsFields } = require("~airtable/tables/paymentRequestsSchema");
+const { createDonorPayment } = require("~airtable/tables/donorPayments");
+
+const { donorPaymentsFields } = require("~airtable/tables/donorPaymentsSchema")
 
 module.exports.register = function register(slackEvents) {
   slackEvents.on("message", filterAndReply);
@@ -18,17 +15,31 @@ const filterAndReply = async (event) => {
   if (event.bot_id) {
     return;
   }
-  const reimbursementChannel = await findChannelByName(REIMBURSEMENT_CHANNEL);
-  if (!isReimbursementReply(event, reimbursementChannel.id)) {
+
+  let reimbursementChannel;
+
+  try {
+    reimbursementChannel = await findChannelByName(REIMBURSEMENT_CHANNEL);
+    if (!isReimbursementReply(event, reimbursementChannel.id)) {
+      return;
+    }
+  } catch (e) {
+    console.error("Error finding reimbursement channel", e);
     return;
   }
-  const [paymentRequest] = await findPaymentRequestBySlackThreadId(
-    event.thread_ts
-  );
-  if (!paymentRequest) {
-    console.log("paymentRequest not posted by bot.");
+
+  let paymentRequest
+  try {
+    [paymentRequest] = await findPaymentRequestBySlackThreadId(event.thread_ts);
+    if (!paymentRequest) {
+      console.log("paymentRequest not posted by bot.");
+      return;
+    }
+  } catch (e) {
+    console.error("Error fetching payment request by Slack thread ID", e);
     return;
   }
+
   const amountMatches = findAmountsInString(event.text);
   if (!amountMatches || amountMatches.length === 0) {
     console.log(`No money matches found in: ${event.text}`);
@@ -37,18 +48,24 @@ const filterAndReply = async (event) => {
 
   const newDonationAmount = Number(amountMatches[0]);
 
-  const [record] = await createDonorPayment({
-    [donorPaymentsFields.amount]: newDonationAmount,
-    [donorPaymentsFields.paymentRequest]: [paymentRequest.getId()],
-    [donorPaymentsFields.status]: donorPaymentsFields.status_options.pending,
-    [donorPaymentsFields.donorSlackId]: event.user,
-    [donorPaymentsFields.recipientConfirmation]:
+  try {
+    const [record] = await createDonorPayment({
+      [donorPaymentsFields.amount]: newDonationAmount,
+      [donorPaymentsFields.paymentRequest]: [paymentRequest.getId()],
+      [donorPaymentsFields.status]: donorPaymentsFields.status_options.pending,
+      [donorPaymentsFields.donorSlackId]: event.user,
+      [donorPaymentsFields.recipientConfirmation]:
       donorPaymentsFields.recipientConfirmation_options.pending,
-    [donorPaymentsFields.donorConfirmation]:
+      [donorPaymentsFields.donorConfirmation]:
       donorPaymentsFields.donorConfirmation_options.confirmed,
-  });
-  if (!record) {
-    console.log("Couldn't add donor's payment");
+    });
+
+    if (!record) {
+      console.log("Couldn't add donor's payment");
+      return;
+    }
+  } catch (e) {
+    console.error("Error adding donor payment", e);
     return;
   }
 
@@ -58,15 +75,21 @@ const filterAndReply = async (event) => {
     newBalance <= 0
       ? "reimbursement is complete!!"
       : `just ${newBalance.toFixed(2)} to go!`;
-  await slackapi.chat.postMessage({
-    token: process.env.SLACK_BOT_TOKEN,
-    channel: reimbursementChannel.id,
-    thread_ts: event.thread_ts,
-    text: `Thanks <@${event.user}>!
-They sent ${newDonationAmount.toFixed(2)}, ${message}`,
-  });
+  
+  try {
+    await slackapi.chat.postMessage({
+      token: process.env.SLACK_BOT_TOKEN,
+      channel: reimbursementChannel.id,
+      thread_ts: event.thread_ts,
+      text: `Thanks <@${event.user}>!
+      They sent ${newDonationAmount.toFixed(2)}, ${message}`,
+    });
+  } catch (e) {
+    console.error("Error posting reimbursement message", e);
+    return;
+  }
 };
-
+  
 const findAmountsInString = (text) => {
   return text.replace(/<.{0,15}>/g, "").match(/\d+\.?\d*/g);
 };
@@ -74,3 +97,5 @@ const findAmountsInString = (text) => {
 const isReimbursementReply = (event, reimbursementChannelId) => {
   return event.channel === reimbursementChannelId && event.parent_user_id;
 };
+
+module.exports.filterAndReply = filterAndReply;
